refactor: rename checkSystemConfig command
